+++ apps/web/src/app/partner/_shared/api/get-review-list.ts
@@ -1,8 +1,8 @@
 import { API_BASE_URL } from "@/shared/api/config";
 import { fetchJson } from "@/shared/api/fetch-json";
 import type { ReviewListFilter } from "@/shared/api/query-keys";
-import { reviewListSchema } from "../../_shared/model/review-list.schema";
-import type { ReviewList } from "../../_shared/model/types";
+import { reviewListSchema } from "../model/review-list.schema";
+import type { ReviewList } from "../model/types";
 
 export function getReviewList(filter?: ReviewListFilter): Promise<ReviewList> {
   const params = new URLSearchParams();
